Add pyarrow.lib.is_opentelemetry_enabled and add info on pa.show_info()
