+++ python/pyarrow/src/arrow/python/helpers.cc
@@ -31,7 +31,6 @@
 #include "arrow/python/decimal.h"
 #include "arrow/type_fwd.h"
 #include "arrow/util/checked_cast.h"
-#include "arrow/util/config.h"
 #include "arrow/util/float16.h"
 #include "arrow/util/logging.h"
 
@@ -489,14 +488,6 @@ Status IntegerScalarToFloat32Safe(PyObject* obj, float* out) {
 void DebugPrint(PyObject* obj) {
   std::string repr = PyObject_StdStringRepr(obj);
   PySys_WriteStderr("%s\n", repr.c_str());
-}
-
-bool IsThreadingEnabled() {
-#ifdef ARROW_ENABLE_THREADING
-  return true;
-#else
-  return false;
-#endif
 }
 
 }  // namespace internal
